+++ tests/conftest.py
@@ -1,0 +1,3 @@
+# coding: utf-8
+
+pytest_plugins = 'tests.pytest_random'