+++ tests/test_encoding.py
@@ -75,13 +75,11 @@ class TestEncoding(UnitTestCase):
         filename = u'espace\xa0 et TM\u2122.doc'
         self.local.make_file('/', filename)
         self.wait_sync(wait_for_async=True)
-
         info = self.remote.get_info('/' + filename)
         assert info.name == filename
 
-    @pytest.mark.skipif(
-        AbstractOSIntegration.is_mac(),
-        reason="Normalization doesn't work on Mac")
+    @pytest.mark.skipif(AbstractOSIntegration.is_mac(),
+                        reason='Normalization dont work on Mac')
     def test_fileinfo_normalization(self):
         self.engine_1.stop()
         name = u'Teste\u0301'