+++ tests/test_local_copy_paste.py
@@ -2,7 +2,9 @@
 import os
 import shutil
 
-from .common_unit_test import FILE_CONTENT, RandomBug, UnitTestCase
+import pytest
+
+from .common_unit_test import FILE_CONTENT, UnitTestCase
 
 
 class TestLocalCopyPaste(UnitTestCase):
@@ -82,8 +84,8 @@ class TestLocalCopyPaste(UnitTestCase):
         assert (len(remote.get_fs_children(self.remote_ref_1))
                 == self.NUMBER_OF_LOCAL_FILES_TOTAL)
 
-    @RandomBug('NXDRIVE-815', target='mac', repeat=5)
-    @RandomBug('NXDRIVE-815', target='windows', repeat=5)
+    @pytest.mark.randombug(
+        'NXDRIVE-815', condition=(sys.platform != 'linux2'), repeat=5)
     def test_local_copy_paste_files(self):
         self._local_copy_paste_files()