+++ tests/test_synchronization.py
@@ -1,16 +1,18 @@
 # coding: utf-8
 import socket
+import sys
 import time
 
 from nuxeo.exceptions import HTTPError
 from requests import ConnectionError
 
+import pytest
+
 from nxdrive.client import LocalClient
 from nxdrive.osi import AbstractOSIntegration
 from . import RemoteTest
 from .common import OS_STAT_MTIME_RESOLUTION, TEST_WORKSPACE_PATH
-from .common_unit_test import (DEFAULT_WAIT_SYNC_TIMEOUT, RandomBug,
-                               UnitTestCase)
+from .common_unit_test import DEFAULT_WAIT_SYNC_TIMEOUT, UnitTestCase
 
 
 class TestSynchronization(UnitTestCase):
@@ -721,7 +723,8 @@ class TestSynchronization(UnitTestCase):
         self.wait_sync(wait_for_async=True)
         assert not local.exists('/test.odt')
 
-    @RandomBug('NXDRIVE-808', target='mac', repeat=2)
+    @pytest.mark.randombug(
+        'NXDRIVE-808', condition=(sys.platform == 'darwin'), repeat=2)
     def test_synchronize_deletion(self):
         local = self.local_1
         remote = self.remote_document_client_1