+++ tests/test_synchronization_suspend.py
@@ -1,10 +1,16 @@
 # coding: utf-8
+import sys
+
+import pytest
+
 from nxdrive.osi import AbstractOSIntegration
-from .common_unit_test import RandomBug, UnitTestCase
+from .common_unit_test import UnitTestCase
 
 
 class TestSynchronizationSuspend(UnitTestCase):
-    @RandomBug('NXDRIVE-805', target='windows', repeat=2)
+
+    @pytest.mark.randombug(
+        'NXDRIVE-805', condition=(sys.platform == 'win32'), repeat=2)
     def test_basic_synchronization_suspend(self):
         local = self.local_1
         remote = self.remote_document_client_1
@@ -67,7 +73,8 @@ class TestSynchronizationSuspend(UnitTestCase):
         assert len(remote.get_children_info(self.workspace_1)) == 2
         assert not engine.get_queue_manager().is_paused()
 
-    @RandomBug('NXDRIVE-812', target='linux', mode='BYPASS')
+    @pytest.mark.randombug(
+        'NXDRIVE-812', condition=(sys.platform == 'linux2'), mode='BYPASS')
     def test_synchronization_end_with_children_ignore_parent(self):
         """ NXDRIVE-655: children of ignored folder are not ignored. """