+++ tests/test_watchers.py
@@ -6,7 +6,7 @@ import pytest
 
 from nxdrive.client import LocalClient
 from nxdrive.osi import AbstractOSIntegration
-from .common_unit_test import RandomBug, UnitTestCase
+from .common_unit_test import UnitTestCase
 
 
 class TestWatchers(UnitTestCase):
@@ -29,7 +29,8 @@ class TestWatchers(UnitTestCase):
         # With root
         assert len(res) == folders + files + 1
 
-    @RandomBug('NXDRIVE-808', target='windows', repeat=2)
+    @pytest.mark.randombug(
+        'NXDRIVE-808', condition=(sys.platform == 'win32'), repeat=2)
     def test_reconcile_scan(self):
         files, folders = self.make_local_tree()
         self.make_server_tree()
@@ -70,8 +71,10 @@ class TestWatchers(UnitTestCase):
         # With root
         assert len(res) == folders + files + 1
 
-    @RandomBug('NXDRIVE-806', target='linux', mode='BYPASS')
-    @RandomBug('NXDRIVE-806', target='windows', repeat=2)
+    @pytest.mark.randombug(
+        'NXDRIVE-806', condition=(sys.platform == 'linux2'), mode='BYPASS')
+    @pytest.mark.randombug(
+        'NXDRIVE-806', condition=(sys.platform == 'win32'), repeat=2)
     def test_local_watchdog_creation(self):
         # Test the creation after first local scan
         self.queue_manager_1.suspend()
@@ -248,9 +251,8 @@ class TestWatchers(UnitTestCase):
         assert not remote.exists(
             u'/P\xf4le applicatif/e\u0302tre ou ne pas \xeatre.odt')
 
-    @pytest.mark.skipif(
-        AbstractOSIntegration.is_windows(),
-        reason='Windows cannot have file ending with a space.')
+    @pytest.mark.skipif(AbstractOSIntegration.is_windows(),
+                        reason='Windows cannot have file ending with a space.')
     def test_watchdog_space_remover(self):
         """
         Test files and folders ending with space.
@@ -273,7 +275,8 @@ class TestWatchers(UnitTestCase):
         assert (remote.get_info(u'/Accentue\u0301.odt').name
                 == u'Accentu\xe9 avec un \xea et un \xe9.odt')
 
-    @RandomBug('NXDRIVE-808', target='mac', repeat=5)
+    @pytest.mark.randombug(
+        'NXDRIVE-808', condition=(sys.platform == 'darwin'), repeat=5)
     def test_watchdog_encoding(self):
         local = self.local_1
         remote = self.remote_document_client_1
@@ -330,7 +333,8 @@ class TestWatchers(UnitTestCase):
         assert not remote.exists(
             u'/Sub folder/e\u0302tre ou ne pas \xeatre.odt')
 
-    @RandomBug('NXDRIVE-808', target='windows', repeat=2)
+    @pytest.mark.randombug(
+        'NXDRIVE-808', condition=(sys.platform == 'win32'), repeat=2)
     def test_watcher_remote_id_setter(self):
         local = self.local_1
         # As some user can rewrite same file for no reason